Change platform of GNOME rules
Rules that are related to GNOME and dconf were marked with the `machine`
platform. But, there isn't anything special about them that would
require to mark them with the `machine` platform. Also, the group
is marked with `package[gdm]` platform which already prevents their
evaluation on the system that don't contain GNOME.

We will remove the `machine` platform. This will make the rules
applicable and passing during a build of a bootable container image.
The `Containerfile` needs to run `dnf install -y gdm` before running
`oscap-im` to see the effect of this commit.

Related to: https://issues.redhat.com/browse/OPENSCAP-5328

test: checking with rule deletion

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

test: checking with rule deletion

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>